GitHub Actions: caching on macOS and other small improvements (#10)
This commit introduces caching for Python and PyInstaller on macOS. This saves anywhere from 5-30 minutes per build (GHA's speed varies wildly based on time of day). Also, a few other small changes, such as using `make -j2` when we experience a cache miss and have to build Python manually, and moving the "if" rules from the bottom to near the top of each step definition.