Open file dialog: default to current level's folder
Minor detail, but it's bothered me for literally years. The "File -> Open Level from File..." dialog will now *always* initially show the folder containing the currently-loaded level, if such a folder exists. (The previous behavior was to show whatever had been most recently selected with a previous QFileDialog in the same session, or just the current working directory if it's the first QFileDialog since launching Reggie.)